lambda_api_fastapi_instrumented_xray/src/requirements.txt
boto3==1.17.*
fastapi==0.67.0
mangum==0.12.1
opentelemetry-instrumentation-fastapi==0.26b1
opentelemetry-api==1.7.1
opentelemetry-sdk-extension-aws==2.0.1